add tiling detection test
